primitives: move encoding tests to integration tests
Move encoding-related tests into `primitives/tests/encoding.rs`
as suggested by the FIXME comment in the codebase.

A new `helpers.rs` file was added in the `tests/` directory to host
shared helper
